feat(router): router infra and home page
- move app to HomePage
- put routes to app
- add BrowserRouter to index
- routes constants
