+++ src/components/index.js
@@ -1,0 +1,9 @@
+export { App } from "./app/app";
+export { AppHeader } from "./app-header/app-header";
+export { BurgerConstructor } from "./burger-constructor/burger-constructor";
+export { BurgerIngredients } from "./burger-ingredients/burger-ingredients";
+export { IngredientDetails } from "./ingredient-details/ingredient-details";
+export { ModalConnectError } from "./modal-connect-error/modal-connect-error";
+export { ModalLoading } from "./modal-loading/modal-loading";
+export { ModalPortal } from "./modal-portal/modal-portal";
+export { OrderDetails } from "./order-details/order-details";
