+++ src/utils/index.js
@@ -1,4 +1,5 @@
 import { STATUSES } from "./connection-statuses";
 import { INGREDIENT_TYPES } from "./ingredients-groups-priority";
+export { ROUTES } from "./routes";
 
 export { STATUSES, INGREDIENT_TYPES };
