+++ src/utils/routes.js
@@ -1,0 +1,3 @@
+export const ROUTES = {
+  ROOT: "/",
+};
